weatherunderground: Adapt to the latest API
The endpoints and parameter changes for now, this itself is a mess. They should have just scrapped the entire thing, rather than releasing whatever this is.

For instance, the current condition endpoint has no info on the observation location but the place id is too long make any sense to the user

The forecast endpoint has forecast for five days but they lack any icon data to map to a weather icon, also the current days forecast values will be made null by early evening.

Now the forecast should almost make sense to the user, while not letting the app crash