Remove now obsolete code
The API seems now uses a iconcode which is same as that is used by the icon packs, so the values can be directly assigned to conditionCode field.

The humidity is now available as Integer which makes the % stripping function useless.

The API now uses language parameter as same as the format returned by android which makes the mapping of the system locale and the language obsolete.

Remove all these code.